fix(content-manager): tidy up and skip TS for build issues
