+++ packages/core/admin/admin/src/content-manager/components/Relations/utils/tests/normalizeRelations.test.ts
@@ -1,3 +1,4 @@
+// @ts-nocheck TODO remove after FE relations changes
 import { normalizeRelations } from '../normalizeRelations';
 
 const FIXTURE_RELATIONS = [
